refactor: refactor types folder structure
