+++ src/app/messenger/types/index.ts
@@ -1,14 +1,14 @@
-export * from './group.model';
-export * from './delete-group-error.model';
-export * from './get-groups-error.model';
-export * from './groups-state.model';
-export * from './post-group-error';
-export * from './users-state.model';
-export * from './post-group-response.model';
-export * from './person.model';
-export * from './get-persons-response.model';
-export * from './get-persons-error.model';
-export * from './get-conversations-response.model';
-export * from './get-conversations-error.model';
-export * from './post-conversation-response.model';
-export * from './post-conversation-error.model';
+export * from './groups/group.model';
+export * from './groups/delete-group-error.model';
+export * from './groups/get-groups-error.model';
+export * from './groups/groups-state.model';
+export * from './groups/post-group-error';
+export * from './users/users-state.model';
+export * from './users/post-group-response.model';
+export * from './users/person.model';
+export * from './users/get-persons-response.model';
+export * from './users/get-persons-error.model';
+export * from './users/get-conversations-response.model';
+export * from './users/get-conversations-error.model';
+export * from './users/post-conversation-response.model';
+export * from './users/post-conversation-error.model';
