+++ src/app/messenger/types/users/get-conversations-response.model.ts
